+++ src/OpenTelemetry.Api/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* `B3Propagator` class from `OpenTelemetry.Extensions.Propagators` namespace has
+  been deprecated and moved as is to a new `OpenTelemetry.Extensions.Propagators`
+  namespace, shipped as part of the `OpenTelemetry.Extensions.Propagators` package.
+  It will be removed in the next major release, see issue [#3259](https://github.com/open-telemetry/opentelemetry-dotnet/issues/3259)
+
 ## 1.3.0-beta.2
 
 Released 2022-May-16